Refactored agents into a module with separate files per agent.
