+++ simulation/agents/__init__.py
@@ -1,0 +1,4 @@
+from agents.marketplayer import MarketPlayer
+from agents.arbitrageur import Arbitrageur
+from agents.banker import Banker
+from agents.randomizer import Randomizer
