+++ simulation/agents/marketplayer.py
@@ -1,12 +1,9 @@
-"""agents.py: Individual agents that will interact with the Havven market."""
-from typing import Set, Tuple, Optional
-import random
+from typing import Set
 
 from mesa import Agent
 
 import orderbook as ob
 import model
-
 
 class MarketPlayer(Agent):
     """
@@ -219,181 +216,3 @@ class MarketPlayer(Agent):
 
     def step(self) -> None:
         pass
-
-class Banker(MarketPlayer):
-    """Wants to buy curits and issue nomins, in order to accrue fees."""
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        self.fiat_curit_order: Optional["ob.Bid"] = None
-        self.nomin_curit_order: Optional["ob.Bid"] = None
-        self.rate: float = random.random() * 0.05
-
-    def step(self) -> None:
-        if self.fiat > 0:
-            if self.fiat_curit_order:
-                self.fiat_curit_order.cancel()
-            fiat = self.model.max_transferrable_fiat(self.fiat)
-            self.fiat_curit_order = self.sell_fiat_for_curits(fiat * self.rate)
-
-        if self.nomins > 0:
-            if self.nomin_curit_order:
-                self.nomin_curit_order.cancel()
-            nomins = self.model.max_transferrable_nomins(self.nomins)
-            self.nomin_curit_order = self.sell_nomins_for_curits(nomins)
-
-        if self.curits > 0:
-            self.escrow_curits(self.curits)
-
-        issuable = self.max_issuance_rights() - self.issued_nomins
-        if issuable > 0:
-            self.issue_nomins(issuable)
-
-
-class Arbitrageur(MarketPlayer):
-    """Wants to find arbitrage cycles and exploit them to equalise prices."""
-
-    def step(self) -> None:
-        """Find an exploitable arbitrage cycle."""
-        # The only cycles that exist are CUR -> FIAT -> NOM -> CUR,
-        # its rotations, and the reverse cycles.
-        # The bot will act to place orders in all markets at once,
-        # if there is an arbitrage opportunity, taking into account
-        # the fee rates.
-
-        if self._forward_multiple_() > 1.1:
-            # Trade in the forward direction
-            # TODO: work out which rotation of this cycle would be the least wasteful
-            # cur -> fiat -> nom -> cur
-            fn_price = 1.0 / self.model.nom_fiat_market.lowest_ask_price()
-            nc_price = 1.0 / self.model.cur_nom_market.lowest_ask_price()
-
-            cf_qty = sum(b.quantity for b in self.model.cur_fiat_market.highest_bids())
-            fn_qty = sum(a.quantity for a in self.model.nom_fiat_market.lowest_asks())
-            nc_qty = sum(a.quantity for a in self.model.cur_nom_market.lowest_asks())
-
-            c_qty = min(self.curits, cf_qty)
-            self.sell_curits_for_fiat(c_qty)
-
-            f_qty = min(self.fiat, fn_qty * fn_price)
-            self.sell_fiat_for_curits(f_qty)
-
-            n_qty = min(self.nomins, nc_qty * nc_price)
-            self.sell_nomins_for_curits(n_qty)
-
-        elif self._reverse_multiple_() > 1.1:
-            # Trade in the reverse direction
-            # cur -> nom -> fiat -> cur
-            fc_price = 1.0 / self.model.cur_fiat_market.lowest_ask_price()
-
-            cn_qty = sum(b.quantity for b in self.model.cur_nom_market.highest_bids())
-            nf_qty = sum(b.quantity for b in self.model.nom_fiat_market.highest_bids())
-            fc_qty = sum(a.quantity for a in self.model.cur_fiat_market.lowest_asks())
-
-            c_qty = min(self.curits, cn_qty)
-            self.sell_curits_for_nomins(c_qty)
-
-            n_qty = min(self.nomins, nf_qty)
-            self.sell_nomins_for_fiat(n_qty)
-
-            f_qty = min(self.fiat, fc_qty * fc_price)
-            self.sell_nomins_for_curits(n_qty)
-
-    def _cycle_fee_rate_(self) -> float:
-        """Divide by this fee rate to determine losses after one traversal of an arbitrage cycle."""
-        return (1 + self.model.nom_transfer_fee_rate) * \
-               (1 + self.model.cur_transfer_fee_rate) * \
-               (1 + self.model.fiat_transfer_fee_rate)
-
-    def _forward_multiple_no_fees_(self) -> float:
-        """
-        The value multiple after one forward arbitrage cycle, neglecting fees.
-        """
-        # cur -> fiat -> nom -> cur
-        return self.model.cur_fiat_market.highest_bid_price() / \
-               (self.model.nom_fiat_market.lowest_ask_price() * self.model.cur_nom_market.lowest_ask_price())
-
-    def _reverse_multiple_no_fees_(self) -> float:
-        """
-        The value multiple after one reverse arbitrage cycle, neglecting fees.
-        """
-        # cur -> nom -> fiat -> cur
-        return self.model.cur_nom_market.highest_bid_price() * self.model.nom_fiat_market.highest_bid_price() / \
-               self.model.cur_fiat_market.lowest_ask_price()
-
-    def _forward_multiple_(self) -> float:
-        """The return after one forward arbitrage cycle."""
-        # Note, this only works because the fees are purely multiplicative.
-        return self._forward_multiple_no_fees_() / self._cycle_fee_rate_()
-
-    def _reverse_multiple_(self) -> float:
-        """The return after one reverse arbitrage cycle."""
-        # As above. If the fees were not just levied as percentages this would need to be updated.
-        return self._reverse_multiple_no_fees_() / self._cycle_fee_rate_()
-
-    def _equalise_tokens_(self) -> None:
-        pass
-
-class Randomizer(MarketPlayer):
-    """Places random bids and asks near current market prices."""
-
-    def __init__(self, unique_id: int, havven: "model.Havven",
-                 fiat: float = 0.0, curits: float = 0.0,
-                 nomins: float = 0.0,
-                 variance: float = 0.05, order_lifetime: int = 30) -> None:
-        super().__init__(unique_id, havven, fiat, curits, nomins)
-        self.variance = variance
-        """This agent will place orders within (+/-)variance*price of the going rate."""
-
-        self.order_lifetime = order_lifetime
-        """Orders older than this lifetime will be cancelled."""
-
-    def step(self) -> None:
-        condemned = []
-        for order in self.orders:
-            if order.book.time > order.time + self.order_lifetime:
-                condemned.append(order)
-        for order in condemned:
-            order.cancel()
-
-        action = random.choice([self._cur_fiat_bid_, self._cur_fiat_ask_,
-                                self._nom_fiat_bid_, self._nom_fiat_ask_,
-                                self._cur_nom_bid_, self._cur_nom_ask_])
-
-        action()
-
-    def _cur_fiat_bid_(self) -> None:
-        price = self.model.cur_fiat_market.price
-        movement = round((2*random.random() - 1) * \
-                         price * self.variance, 3)
-        self.place_curits_fiat_bid(self.fiat/10, price + movement)
-
-    def _cur_fiat_ask_(self) -> None:
-        price = self.model.cur_fiat_market.price
-        movement = round((2*random.random() - 1) * \
-                         price * self.variance, 3)
-        self.place_curits_fiat_ask(self.fiat/10, price + movement)
-
-    def _nom_fiat_bid_(self) -> None:
-        price = self.model.cur_fiat_market.price
-        movement = round((2*random.random() - 1) * \
-                         price * self.variance, 3)
-        self.place_nomins_fiat_bid(self.fiat/10, price + movement)
-
-    def _nom_fiat_ask_(self) -> None:
-        price = self.model.cur_fiat_market.price
-        movement = round((2*random.random() - 1) * \
-                         price * self.variance, 3)
-        self.place_nomins_fiat_ask(self.fiat/10, price + movement)
-
-    def _cur_nom_bid_(self) -> None:
-        price = self.model.cur_fiat_market.price
-        movement = round((2*random.random() - 1) * \
-                         price * self.variance, 3)
-        self.place_curits_nomins_bid(self.fiat/10, price + movement)
-
-    def _cur_nom_ask_(self) -> None:
-        price = self.model.cur_fiat_market.price
-        movement = round((2*random.random() - 1) * \
-                         price * self.variance, 3)
-        self.place_curits_nomins_ask(self.fiat/10, price + movement)
